orange: Rename PurpleGainDeletion to GeneCopyNumber
