--- orange-datamodel/src/main/java/com/hartwig/hmftools/datamodel/finding/CopyNumber.java
@@ -1,15 +1,0 @@
-package com.hartwig.hmftools.datamodel.finding;
-
-import com.hartwig.hmftools.datamodel.driver.Driver;
-import com.hartwig.hmftools.datamodel.purple.PurpleGainDeletion;
-
-import org.immutables.gson.Gson;
-import org.immutables.value.Value;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-@Gson.TypeAdapters
-@Value.Immutable
-@Value.Style(passAnnotations = { NotNull.class, Nullable.class })
-public interface CopyNumber extends Driver, PurpleGainDeletion {
-}
