+++ orange-datamodel/src/main/java/com/hartwig/hmftools/datamodel/finding/GainDeletion.java
@@ -1,7 +1,7 @@
-package com.hartwig.hmftools.datamodel.purple;
+package com.hartwig.hmftools.datamodel.finding;
 
-import com.hartwig.hmftools.datamodel.driver.DriverInterpretation;
-import com.hartwig.hmftools.datamodel.driver.ReportedStatus;
+import com.hartwig.hmftools.datamodel.driver.Driver;
+import com.hartwig.hmftools.datamodel.purple.CopyNumberInterpretation;
 
 import org.immutables.gson.Gson;
 import org.immutables.value.Value;
@@ -11,8 +11,8 @@ import org.jetbrains.annotations.Nullable;
 @Gson.TypeAdapters
 @Value.Immutable
 @Value.Style(passAnnotations = { NotNull.class, Nullable.class })
-public interface PurpleGainDeletion
-{
+public interface GainDeletion extends Driver {
+
     @NotNull
     CopyNumberInterpretation interpretation();
 
@@ -33,10 +33,4 @@ public interface PurpleGainDeletion
     double minCopies();
 
     double maxCopies();
-
-    @NotNull
-    ReportedStatus reportedStatus();
-
-    @NotNull
-    DriverInterpretation driverInterpretation();
 }
